view handler
This adds a special view handler mixin that forces the content type
to `:json` and updates `halt` to return an empty json object body if
no custom body is specified.

The goal here is to make working with json in view handlers easier.
Some http clients require that the body be valid json if the response
content type is json (or even if the request made requested json).
Halting by default returns an empty body which isn't valid json.