view handler: update `halt`to not break flexible args behavior
This underlying `halt` on a deas' view handler allows for multiple
combinations of arguments to be passes with fallbacks to intelligent
defaults.  The previous implementation essentially forced sending
args in a specific order.  This mean you couldn't call `halt` like
`halt(404, 'unknown resource')` using a json view handler.

This was not the original intent.  The intent is that you can use
`halt` just as you would on a normal view handler.  This change
preserves that original intent, with fallbacks to json-intelligent
defaults for args that are not specified.