bring in much-plugin; move init behavior into a before init hook
In deas view handlers, it is better to *not* override the initializers
and instead use the callback system to attach init behavior.  However,
this is dangerous to do in included methods as that every time the
mixin is included (even on a receiver it was already included on)
this method will execute.  This leaves the possibility of adding
multiple of the same deas view handler callbacks.

This brings in the much-plugin gem when ensures the plugin included
procs are only executed once per receiver and then replaces the
initializer override with a before init callback.

There are no behavior or test changes - just a cleanup to more
properly bind this behavior to deas json view handlers.