call the `body` helper when halting
This ensures that whatever body value we halt with will adhere to
Rack's spec in the response.  This involved also forcing the
latest Deas version (0.43.4) which similarly ensures halt bodies
adhere to Rack's spec.

I also changed the check for whether to send the default body or
not to be based on only if it is given an "empty" body.  I chose
to just test for super obvious stuff for now: `nil` or `""`.  The
goal here is to always return something that is valid json if at
all possible.