don't call Deas' body helper when halting
Deas already handles this for you in its `halt` method.  This
supers to that method so it takes care of it for us.  This reverts
part of the work in PR 6.

I noticed this when I went to put Deas-Json in one of our apps and
saw errors come up in cases where we set a custom body using `body`
before halting.  I've updated the test suite to account for this
case as well.