switch to using Deas' new default response handler API
This is Deas' new way of letting handlers override the default
response values on a per-handler basis.  Now, Deas-Json's view
handler defaults the status and body and this applies whether
`halt` is used or not.  This is a more formal, thorough, and
universal way for Deas-Json to change the defaults and ensure they
are used in the absence of any instance specified status/body.

The goal of doing this is unchanged: to provide a valid empty
json object body if no formal body is specified so that browsers
and client side js doesn't error trying to parse json responses.
This only changes the manner used to accomplish this goal.

Note: this switched to adding some system tests to further and
more thoroughly test that responses were being defaulted correctly.
With the simpler handler implementation, its unit tests felt a
little wanting.